requirements.txt
absl-py==0.3.0
astor==0.7.1
certifi==2018.4.16
chardet==3.0.4
click==6.7
Django==2.0.6
djangorestframework==3.8.2
Flask==1.0.2
gast==0.2.0
grpcio==1.14.1
idna==2.7
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
numpy==1.14.5
pkg-resources==0.0.0
protobuf==3.6.0
pyee==5.0.0
pytz==2018.4
requests==2.19.1
six==1.11.0
slackclient==1.2.1
slackeventsapi==1.1.0
tensorboard==1.10.0
tensorflow==1.10.0
termcolor==1.1.0
urllib3==1.23
websocket-client==0.48.0
Werkzeug==0.14.1
